+++ gazebo/rendering/ScenePrivate.hh
@@ -329,6 +329,9 @@ namespace gazebo
 
       /// \brief Keep track of data of joints.
       public: JointMsgs_M joints;
+
+      /// \brief Size of shadow texture
+      public: unsigned int shadowTextureSize = 1024u;
     };
   }
 }