+++ Changelog.md
@@ -8,6 +8,9 @@
 1. Support Heightmap LOD
     * [Pull request 2636](https://bitbucket.org/osrf/gazebo/pull-request/2636)
 
+1. Support setting shadow texture size
+    * [Pull request 2644](https://bitbucket.org/osrf/gazebo/pull-request/2644)
+
 ## Gazebo 7.5.0 (2017-01-11)
 
 1. Remove qt4 webkit in gazebo7 (used for HotkeyDialog).